move handler func out of CLI definition code
