+++ smartsim/_core/_cli/clean.py
@@ -25,53 +25,13 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 import argparse
-import shutil
 import typing as t
 
-from smartsim._core._cli.utils import get_install_path, MenuItem
+from smartsim._core._cli.utils import get_install_path, MenuItem, clean
 from smartsim.log import get_logger
 
 smart_logger_format = "[%(name)s] %(levelname)s %(message)s"
 logger = get_logger("Smart", fmt=smart_logger_format)
-
-
-def clean(core_path: str, _all: bool = False) -> None:
-    """Remove pre existing installations of ML runtimes
-
-    :param _all: Remove all non-python dependencies
-    :type _all: bool, optional
-    """
-
-    build_temp = core_path / ".third-party"
-    if build_temp.is_dir():
-        shutil.rmtree(build_temp, ignore_errors=True)
-
-    lib_path = core_path / "lib"
-    if lib_path.is_dir():
-
-        # remove RedisAI
-        rai_path = lib_path / "redisai.so"
-        if rai_path.is_file():
-            rai_path.unlink()
-            logger.info("Successfully removed existing RedisAI installation")
-
-        backend_path = lib_path / "backends"
-        if backend_path.is_dir():
-            shutil.rmtree(backend_path, ignore_errors=True)
-            logger.info("Successfully removed ML runtimes")
-
-    bin_path = core_path / "bin"
-    if bin_path.is_dir() and _all:
-        files_to_remove = ["redis-server", "redis-cli", "keydb-server", "keydb-cli"]
-        removed = False
-        for _file in files_to_remove:
-            file_path = bin_path.joinpath(_file)
-
-            if file_path.is_file():
-                removed = True
-                file_path.unlink()
-        if removed:
-            logger.info("Successfully removed SmartSim database installation")
 
 
 class Clean(MenuItem):
